Conflicts:
	build/scripts/utils.gant

+++ plugins/gradle/src/org/jetbrains/plugins/gradle/service/project/GradleExecutionHelper.java
@@ -25,6 +25,7 @@ import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.util.ArrayUtilRt;
 import com.intellij.util.Function;
 import com.intellij.util.SystemProperties;
+import com.intellij.util.containers.ContainerUtil;
 import com.intellij.util.containers.ContainerUtilRt;
 import org.gradle.StartParameter;
 import org.gradle.tooling.*;
@@ -37,6 +38,7 @@ import org.jetbrains.annotations.Nullable;
 import org.jetbrains.plugins.gradle.settings.DistributionType;
 import org.jetbrains.plugins.gradle.settings.GradleExecutionSettings;
 import org.jetbrains.plugins.gradle.util.GradleConstants;
+import org.jetbrains.plugins.gradle.util.GradleUtil;
 
 import java.io.File;
 import java.io.IOException;
@@ -81,10 +83,12 @@ public class GradleExecutionHelper {
   public BuildLauncher getBuildLauncher(@NotNull final ExternalSystemTaskId id,
                                         @NotNull ProjectConnection connection,
                                         @Nullable GradleExecutionSettings settings,
-                                        @NotNull ExternalSystemTaskNotificationListener listener)
+                                        @NotNull ExternalSystemTaskNotificationListener listener,
+                                        @Nullable final String vmOptions)
   {
     BuildLauncher result = connection.newBuild();
-    prepare(result, id, settings, listener, Collections.<String>emptyList());
+    List<String> extraJvmArgs = vmOptions == null ? ContainerUtil.<String>emptyList() : ContainerUtil.newArrayList(vmOptions.trim());
+    prepare(result, id, settings, listener, extraJvmArgs);
     return result;
   }
 
@@ -181,9 +185,16 @@ public class GradleExecutionHelper {
                                      @NotNull String projectPath,
                                      @NotNull GradleExecutionSettings settings,
                                      @NotNull ExternalSystemTaskNotificationListener listener) {
+
+    if (!settings.getDistributionType().isWrapped()) return;
+
+    if (settings.getDistributionType() == DistributionType.DEFAULT_WRAPPED &&
+        GradleUtil.findDefaultWrapperPropertiesFile(projectPath) != null) {
+      return;
+    }
     ProjectConnection connection = getConnection(projectPath, settings);
     try {
-      BuildLauncher launcher = getBuildLauncher(id, connection, settings, listener);
+      BuildLauncher launcher = getBuildLauncher(id, connection, settings, listener, null);
       try {
         final File tempFile = FileUtil.createTempFile("wrap", ".gradle");
         tempFile.deleteOnExit();
@@ -253,15 +264,17 @@ public class GradleExecutionHelper {
           }
           break;
         case WRAPPED:
-          if(settings.getWrapperPropertyFile() != null) {
+          if (settings.getWrapperPropertyFile() != null) {
             File propertiesFile = new File(settings.getWrapperPropertyFile());
-            Distribution distribution =
-              new DistributionFactoryExt(StartParameter.DEFAULT_GRADLE_USER_HOME).getWrappedDistribution(propertiesFile);
-            try {
-              setField(connector, "distribution", distribution);
-            }
-            catch (Exception e) {
-              throw new ExternalSystemException(e);
+            if (propertiesFile.exists()) {
+              Distribution distribution =
+                new DistributionFactoryExt(StartParameter.DEFAULT_GRADLE_USER_HOME).getWrappedDistribution(propertiesFile);
+              try {
+                setField(connector, "distribution", distribution);
+              }
+              catch (Exception e) {
+                throw new ExternalSystemException(e);
+              }
             }
           }
           break;
